Fix UITests that use File Open dialog to load documents
It relied upon the component being loaded before the next
Python code line is executed. OTOH, UITest.load_file uses
EventListener to wait for the OnLoad event, which should
be reused in all such cases.

This moves the waiting code to a dedicated context manager,
and modifies UITest.load_file to use it. Respective tests
using the File dialog also use the new manager to wrap the
action starting the loading process.

Change-Id: Ic6eb0533b06e0ccc63dffc2ddc0e91d8159cae4a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/116693
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>