tdf#130428 remove unnecessary usage of SfxItemState::UNKNOWN
FontworkBar::getState does not need to check for SfxItemState::UNKNOWN
at all, actions solely depend on FontWork object being selected.
This also greatly simplifies that method. Also, the optimization
by passing in a variable to checkForSelectedFontWork and remember
if already computed can be removed - also in other places where
it had to be given, but was not re-used at all

Change-Id: I35b1f36195feb1d645619665d2dd65a84b75b118
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/117014
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>