tdf#104239 doc import: accept outlineLvl0 == (listLvl = MAXLEVEL)
If listLvl is undefined, it is treated as level 0,
so when testing that outlineLvl is equal to listLvl,
then consider MAXLEVEL to be equivalent to zero.

The opposite is not true. An undefined outlineLvl
is not considered to be level 0.

This doesn't fix bug 104239, but it is one step
in the right direction.

Two unit tests followed this path,
(transparent-text.doc and tdf80635_pageLeft.doc)
but neither one sets a numbering style,
so they don't make good unit tests.

Change-Id: I62cd7f5500fcd6dc06327900c33f12c129610a04
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/117745
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>