tdf#104239 doc import: use "Outline" for m_pChosenWW8OutlineStyle
This looks like it was a logic error.
Probably they were only looking at the situation where
the paragraph style set the numbering style to "Outline".
Well, in this case, Heading 1 is actually not part
of Chapter Numbering, but it was the body sprm
that set the style.

In any case, whenever the LFO chosen as the outline
style is referenced, it should apply LO's special,
unique and extremely troublesome "Outline" numbering style.

Change-Id: I354d9b9f4a48abed3453ee046ddd63cda174f25d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/117746
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>