Resolves: tdf#122191 BOOLEAN is a keyword, treat it as such
... to not end up with "BOOL"E"AN" instead, which is a date type
with an era year and literal strings.

This never worked but only was uncovered by

    commit ce4fc2fc08be8ea2773194e303ed42d2579e93a0
    CommitDate: Fri Mar 2 20:27:45 2018 +0100

        Resolves: tdf#115351 convert boolean equivalent format codes to proper Boolean

if the format also had to be converted between locales.

Also preserve boolean equivalent formats during Excel export and
try hard to convert back as much as possible if a literal boolean
string format is a Boolean equivalent of the target locale.

Change-Id: I54f65c276cbf7bb99e960b6d7053c5fa95fbccb6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/118591
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins