related tdf141964 writerfilter CN: drop obsolete exception
First of all, the presence of an outlineLevel is not
equivalent to being part of Chapter Numbering.
Plus even if it is a chapter numbering style
it should be fine to write "Outline" as the numStyleName
after all of the prior rework of numbering import.
So lets not keep unnecessary exceptions.

Change-Id: I89149e199eddacefd0971c805e03d3ad66a1672b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/118706
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>