ScriptForge - (SF_Base) new CloseFormDocument() method
The closure of a form document is available from
the SF_Form service.

A form document is since recent commits associated with
actions like PrintOut() that do not require to access
its internal forms.

As a consequence, the closure of a form document from
the SF_Base service gains in relevance.

Hence the actual commit.

Change-Id: I9d4af078e6631d18c2d92c17249be6025ac9dfdd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/118857
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>