tdf#143455 exclude diagram from pos and margin adaptions
The current implementation of diagram (SmartArt) uses a SdrGroupObj,
but it is actually not a group of the diagram parts. Currently the
group's object list contains only one, dummy object. Therefore it
works not well with the code parts used for groups.

Because we will have (hopefully) the tender 'Dynamic Diagram feature'
I have done nothing on diagrams, but as workaround only excluded
diagrams from these code parts. That way the rendering is same as
in LO 7.1.

This part of code needs to be touched again, when diagrams are
implemented in a different way.

Change-Id: Ife4bb495c96fb1dd57663f28409dc2bb456282ab
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/119344
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>