sw: use followby parameter of outline in Chapter field
When Chapter field with format "Chapter number and name" is used
it does not insert any delimiter between number and name. Best is
to use outline/list property LabelFollowBy.

But since it is done inside field we are limited with supported
characters, so practically only space and nothing are correctly
supported. Tab and newline are replaced by space.

Change-Id: I4583b7051ae5ad963132980443fa70b5a19354e7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/119428
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>