tdf#131546 DOCX import: fix performance regression at tables
Commit 2ab481b038b62b1ff576ac4d49d03c1798cd7f84 "tdf#90069 DOCX:
fix character style of new table rows" caused ~20% slowing down
in loading time of documents with huge tables, related to the
extra processing of the redundant w:rPr of table paragraph runs.
(In DOCX tables, MSO exports the run properties into the run and
paragraph sections too, probably because of compatibility or
usability reasons.)

Theoretically in this case, the run properties which are under the
run section win. On the other hand, because LO copies the props
which are applied on paragraph level, and only them, when copying
a row (e.g. upon inserting a new one), it was needed to apply the
mentioned run props not only as direct character formatting, but
as a direct paragraph formatting too. This way, the support of
copying of rows are solved. Unfortunately, this "double" applying
was done for every single paragraph, which quite slowed down the
opening time. This patch gives a workaround, which completely removes
this double applying functionality in the writerfilter by reverting
commit 2ab481b038b62b1ff576ac4d49d03c1798cd7f84 (except its unit test),
and copy the mentioned run properties into paragraph level, when its
needed: upon inserting a new row before/after. This way we spare a lot
of cycles, as most of the original applies had no real use whatsoever.

Change-Id: Ic24d0d1830e9ec43323e5e213d6481ec28e9abe3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/118985
Tested-by: László Németh <nemeth@numbertext.org>
Reviewed-by: László Németh <nemeth@numbertext.org>