sw bibiliography: limit browse button / page number widgets to the local URL
The normal URL is typically a HTTP resource, so the browse button is not
helpful there after all. And if the URL is copied in manually anyway,
then the dedicated page number widgets are not that helpful, either. So
remove all this, and only keep it for local URLs only.

Change-Id: I25733810cf38981c50014e06201555d98fd9f77f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/120420
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins